pyproject.toml
[tool.poetry]
name = "ocr-ai"
version = "0.1.0"
description = ""
authors = ["Hammad-Rehman <124163466+developer-hammad-rehman@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-jose = "^3.3.0"
psycopg2 = "^2.9.9"
sqlmodel = "^0.0.19"
openai = "1.33.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pypdf2 = {extras = ["crypto"], version = "^3.0.1"}
pandas = "^2.2.2"
pymupdf = "^1.24.5"
httpx = "^0.27.0"
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
